requirements.txt
amqp==5.1.1
asgiref==3.5.2
asttokens==2.0.8
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Deprecated==1.2.13
Django==4.0.4
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.21.3
executing==0.10.0
h11==0.13.0
idna==3.3
inflection==0.5.1
ipdb==0.13.9
ipython==8.4.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.14.0
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
openapi-codec==1.3.2
outcome==1.2.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.4.3
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stack-data==0.4.0
swagger-spec-validator==2.7.4
toml==0.10.2
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
wsproto==1.1.0
